# src/main/java/com/cqjtu/modules/sys/controller/SysRoleController.java
package com.cqjtu.modules.sys.controller;

import com.cqjtu.common.annotation.SysLog;
import com.cqjtu.common.utils.PageUtils;
import com.cqjtu.common.utils.R;
import com.cqjtu.common.validator.ValidatorUtils;
import com.cqjtu.modules.sys.entity.SysRoleEntity;
import com.cqjtu.modules.sys.service.ISysRoleDeptService;
import com.cqjtu.modules.sys.service.ISysRoleMenuService;
import com.cqjtu.modules.sys.service.ISysRoleService;
import org.apache.shiro.authz.annotation.RequiresPermissions;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Map;

/** 角色管理 */
@RestController
@RequestMapping("/sys/role")
public class SysRoleController extends AbstractController {
  @Autowired private ISysRoleService sysRoleService;
  @Autowired private ISysRoleMenuService sysRoleMenuService;
  @Autowired private ISysRoleDeptService sysRoleDeptService;

  /** 角色列表 */
  @RequestMapping("/list")
  @RequiresPermissions("sys:role:list")
  public R list(@RequestParam Map<String, Object> params) {
    PageUtils page = sysRoleService.queryPage(params);

    return R.ok().put("page", page);
  }

  /** 角色列表 */
  @RequestMapping("/select")
  @RequiresPermissions("sys:role:select")
  public R select() {
    List<SysRoleEntity> list = sysRoleService.selectList(null);

    return R.ok().put("list", list);
  }

  /** 角色信息 */
  @RequestMapping("/info/{roleId}")
  @RequiresPermissions("sys:role:info")
  public R info(@PathVariable("roleId") Long roleId) {
    SysRoleEntity role = sysRoleService.selectById(roleId);

    // 查询角色对应的菜单
    List<Long> menuIdList = sysRoleMenuService.queryMenuIdList(roleId);
    role.setMenuIdList(menuIdList);

    // 查询角色对应的部门
    List<Long> deptIdList = sysRoleDeptService.queryDeptIdList(new Long[] {roleId});
    role.setDeptIdList(deptIdList);

    return R.ok().put("role", role);
  }

  /** 保存角色 */
  @SysLog("保存角色")
  @RequestMapping("/save")
  @RequiresPermissions("sys:role:save")
  public R save(@RequestBody SysRoleEntity role) {
    ValidatorUtils.validateEntity(role);

    sysRoleService.save(role);

    return R.ok();
  }

  /** 修改角色 */
  @SysLog("修改角色")
  @RequestMapping("/update")
  @RequiresPermissions("sys:role:update")
  public R update(@RequestBody SysRoleEntity role) {
    ValidatorUtils.validateEntity(role);

    sysRoleService.update(role);

    return R.ok();
  }

  /** 删除角色 */
  @SysLog("删除角色")
  @RequestMapping("/delete")
  @RequiresPermissions("sys:role:delete")
  public R delete(@RequestBody Long[] roleIds) {
    sysRoleService.deleteBatch(roleIds);

    return R.ok();
  }
}